-> Breaking out some of the metadata- and database-specific methods from cohorts/views and into their own helpers file -> Squelching the warning from MySQL about the cursor when fetching metadata value domains
